moves LabeledGraphBuilder to its own file
